simplify local network
